# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: IBooking API
  description: 自习室座位预约系统(IBooking)调用端接口
  contact:
    name: Penglei Mao
    email: penglei.x.mao@gmail.com
  version: 1.0.0
servers:
  - url: http://10.177.44.64/ibooking/
  - url: http://10.177.44.64/ibooking-dev/api/
paths:
...
  /campuses:
    get:
      tags:
        - Campus
      summary: Get an array of all campuses
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'  
    post:   
      tags:
        - Campus
      summary: Create a new campus info
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Campus'
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'       
  /campuses/{campus_id}:
    get:
      tags:
        - Campus
      summary: Get the info of the specified campus
      parameters:
        - name: campus_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    patch:
      tags:
        - Campus
      summary: Update the info of the specified campus
      parameters:
        - name: campus_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Campus'
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'          
    delete:
      tags:
        - Campus
      summary: Delete the info of the specified campus
      parameters:
        - name: campus_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'          
  /campuses/{campus_id}/buildings:
    get:
      tags:
        - Campus
      summary: Get an array of all buildings on the specified campus
      parameters:
        - name: campus_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500' 
  /campuses/name/{campus_name}/buildings:
    get:
      tags:
        - Campus
      summary: Get an array of all buildings on the specified campus
      parameters:
        - name: campus_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'    
  /buildings:
    get:
      tags:
        - Building
      summary: Get an array of all buildings
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'  
    post:   
      tags:
        - Building
      summary: Create a new building info
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Building'
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'            
  /buildings/{building_id}:
    get:
      tags:
        - Building
      summary: Get the info of the specified building
      parameters:
        - name: building_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    patch:
      tags:
        - Building
      summary: Update the info of the specified building
      parameters:
        - name: building_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Building'
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'          
    delete:
      tags:
        - Building
      summary: Delete the info of the specified building
      parameters:
        - name: building_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'          
  /buildings/{building_id}/campus:
    get:
      tags:
        - Building
      summary: Get the campus of the specified building
      parameters:
        - name: building_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/200'
                  - type: object
                    properties:
                      campus:
                        $ref: '#/components/schemas/Campus'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'                
  /buildings/name/{building_name}/campus:
    get:
      tags:
        - Building
      summary: Get an array of all rooms on the specified building
      parameters:
        - name: building_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500' 
  /buildings/{building_id}/rooms:
    get:
      tags:
        - Building
      summary: Get an array of all rooms on the specified building
      parameters:
        - name: building_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /buildings/name/{building_name}/rooms:
    get:
      tags:
        - Building
      summary: Get an array of all rooms on the specified building
      parameters:
        - name: building_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500' 
  /rooms:
    get:
      tags:
        - Room
      summary: Get an array of all rooms
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'  
    post:   
      tags:
        - Room
      summary: Create a new room info
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'            
  /rooms/{room_id}:
    get:
      tags:
        - Room
      summary: Get the info of the specified room
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    patch:
      tags:
        - Room
      summary: Update the info of the specified room
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'    
    delete:
      tags:
        - Room
      summary: Delete the info of the specified room
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'          
  /rooms/{room_id}/building:
    get:
      tags:
        - Room
      summary: Get the building of the specified room
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'           
  /rooms/{room_id}/seats:
    get:
      tags:
        - Room
      summary: Get all seats of the specified room
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /rooms/name/{room_name}/building:
    get:
      tags:
        - Room
      summary: Get the building of the specified room
      parameters:
        - name: room_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'           
  /rooms/name/{room_name}/seats:
    get:
      tags:
        - Room
      summary: Get all seats of the specified room
      parameters:
        - name: room_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /reservations:
    get:
      tags:
        - Reservation
      summary: Get an array of all reservations
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'  
    post:
      tags:
        - Reservation
      summary: Create a new reservation info
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'          
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'          
  /reservations/{reservation_id}:
    get:
      tags:
        - Reservation
      summary: Get the info of the specified reservation
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
            format: int64            
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    patch:
      tags:
        - Reservation
      summary: Update the info of the specified reservation
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
            format: int64            
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - Reservation
      summary: Delete the info of the specified reservation
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'        
  /reservations/{reservation_id}/user:
    get:
      tags:
        - Reservation
      summary: Get the user of the specified reservation
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /reservations/{reservation_id}/seat:
    get:
      tags:
        - Reservation
      summary: Get the seat of the specified reservation
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /reservations/{reservation_id}/signIn:
    get:
      tags:
        - Reservation
      summary: Get the seat of the specified reservation
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
components:
  schemas:
    200:
      type: object
      description: OK
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 请求成功
...
    400:
      type: object
      description: Bad Request
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: 请求格式错误，服务器无法理解该请求
...
    500:
      type: object
      description:  Internal Server Error      
      properties:
        code:
          type: integer
          example: 500
        message:
          type: string
          example: 服务器内部错误
...
    Campus:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: 江湾校区
    Building:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: 智华楼
        alias:
          type: string
          example: JB
        campus:
          $ref: '#/components/schemas/Campus'          
    Room:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: '101'
        building:
          $ref: '#/components/schemas/Building'
        total_capacity:
          type: integer
          format: int64
          example: 50
        power_socket_capacity:
          type: integer
          format: int64
          example: 20
        status:
          type: string
          enum:
            - Open
            - Closed
            - Under maintenance
    Seat:
      type: object
      properties:
        id:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
          example: 1
        power_supply:
          type: boolean
          example: true
        room:
          $ref: '#/components/schemas/Room'
    Reservation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
        seat:
          $ref: '#/components/schemas/Seat'
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - Confirmed
            - Cancelled
            - Pending
        status_change_time:
          type: string
          format: date-time
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 001
        name:
          type: string
          example: penglei
        email:
          type: string
          example: penglei.x.mao@gmail.com
        password:
          type: string
          example: 123456
...
  responses:
...
    400:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/400'
...
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/500'            